update: 将初始化SWDIO部分放在APP区
