+++ projects/HSLink-Pro/src/main.c
@@ -1,5 +1,7 @@
 #include <hpm_romapi.h>
 #include <hpm_dma_mgr.h>
+#include <hpm_gpio_drv.h>
+#include <hpm_gpiom_drv.h>
 #include "board.h"
 #include "dap_main.h"
 #include "WS2812.h"
@@ -27,6 +29,17 @@ static void serial_number_init(void)
     printf("Serial number: %s\n", serial_number);
 }
 
+ATTR_ALWAYS_INLINE
+static inline void SWDIO_DIR_Init(void)
+{
+    HPM_IOC->PAD[SWDIO_DIR].FUNC_CTL = IOC_PAD_FUNC_CTL_ALT_SELECT_SET(0);;
+
+    gpiom_set_pin_controller(HPM_GPIOM, GPIO_GET_PORT_INDEX(SWDIO_DIR), GPIO_GET_PIN_INDEX(SWDIO_DIR), gpiom_core0_fast);
+    gpio_set_pin_output(HPM_FGPIO, GPIO_GET_PORT_INDEX(SWDIO_DIR), GPIO_GET_PIN_INDEX(SWDIO_DIR));
+    gpio_write_pin(HPM_FGPIO, GPIO_GET_PORT_INDEX(SWDIO_DIR), GPIO_GET_PIN_INDEX(SWDIO_DIR), 1);
+}
+
+
 int main(void)
 {
     board_init();
@@ -34,6 +47,8 @@ int main(void)
     board_init_led_pins();
     board_init_usb_pins();
     dma_mgr_init();
+
+    SWDIO_DIR_Init();
 
     printf("version: %s\n", DAP_FW_VER);
     extern char *string_descriptors[];
